% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowinAPI.R
\name{extract_districts}
\alias{extract_districts}
\title{Fetch Districts of a particular state}
\usage{
extract_districts(state_id)
}
\arguments{
\item{state_id}{State ID (get from extract_states() function)}
}
\value{

}
\description{
Fetch Districts of a particular state
}
\examples{
df.District <- extract_districts(16)
}
\author{
Deepanshu Bhalla
}
