% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowinAPI.R
\name{extract_districts_all}
\alias{extract_districts_all}
\title{Fetch Districts of all the Indian states}
\usage{
extract_districts_all(df)
}
\arguments{
\item{df}{Frame Name of output of extract_states() function}
}
\value{

}
\description{
Fetch Districts of all the Indian states
}
\examples{
df.Districts <- extract_districts_all(df.States)
}
\author{
Deepanshu Bhalla
}
