% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowinAPI.R
\name{slots_cleaned}
\alias{slots_cleaned}
\title{Clean Output of slots_district() and slots_pincode()}
\usage{
slots_cleaned(slots.df, age_limit = NULL)
}
\arguments{
\item{slots.df}{Dataframe obtained from slots_district() function}

\item{age_limit}{Enter 18 for 18+, 45 for 45+. By default, combination of both}
}
\value{

}
\description{
Clean Output of slots_district() and slots_pincode()
}
\examples{
slots2 <-  slots_cleaned(slots, age_limit = 45)
}
\author{
Deepanshu Bhalla
}
