% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowinAPI.R
\name{slots_district_all}
\alias{slots_district_all}
\title{Vaccine slots for 7 days from a specific date for multiple districts}
\usage{
slots_district_all(districts, date = Sys.Date(), age_limit = 45)
}
\arguments{
\item{age_limit}{Enter 18 for 18+, 45 for 45+. By default, combination of both}

\item{slots.df}{Dataframe obtained from slots_district() function}
}
\value{

}
\description{
Vaccine slots for 7 days from a specific date for multiple districts
}
\examples{
district.ids <- df.Districts \%>\% select(1) \%>\% pull() \%>\% .[1:2]
slots.all <- slots_district_all(district.ids)
}
\author{
Deepanshu Bhalla
}
