% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowinAPI.R
\name{slots_pincode_all}
\alias{slots_pincode_all}
\title{Vaccine slots for 7 days from a specific date for multiple pincodes}
\usage{
slots_pincode_all(pincodes, date = Sys.Date(), age_limit = 45)
}
\arguments{
\item{date}{slots shown for 7 days from this date}

\item{pincode}{6 digit pincode where you want availability of slots}
}
\value{

}
\description{
Vaccine slots for 7 days from a specific date for multiple pincodes
}
\examples{
slots.all <- slots_pincode_all(pincodes = c(201301,110032))
}
\author{
Deepanshu Bhalla
}
